+++ export_pdf.php
@@ -15,17 +15,40 @@
 // along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
 
 /**
- * Plugin version and other metadata.
+ * PDF export handler.
  *
  * @package    local_student_monitor
  * @copyright  2025 UNCHK - Université Numérique Cheikh Hamidou Kane
  * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
-defined('MOODLE_INTERNAL') || die();
+require_once(__DIR__ . '/../../config.php');
 
-$plugin->component = 'local_student_monitor';
-$plugin->version = 2025111704;  // YYYYMMDDXX
-$plugin->requires = 2022041900; // Moodle 4.0
-$plugin->maturity = MATURITY_STABLE;
-$plugin->release = 'v1.4.0';
+require_login();
+require_capability('local/student_monitor:viewreports', context_system::instance());
+
+$type = required_param('type', PARAM_ALPHA);
+$risklevel = optional_param('risk', '', PARAM_TEXT);
+$startdate = optional_param('startdate', 0, PARAM_INT);
+$enddate = optional_param('enddate', 0, PARAM_INT);
+
+$pdfmanager = new \local_student_monitor\manager\pdf_manager();
+
+switch ($type) {
+    case 'students':
+        $pdfmanager->export_students_pdf($risklevel);
+        break;
+
+    case 'detailed':
+        $pdfmanager->export_detailed_report_pdf();
+        break;
+
+    case 'notifications':
+        $pdfmanager->export_notifications_pdf($startdate, $enddate);
+        break;
+
+    default:
+        print_error('invalidexporttype', 'local_student_monitor');
+}
+
+exit;